Update build-publish-staging.yml
This works.  We're migrating Jenkins and moving away from the old curl method to call the job due to deprecation.  This new action passes the variables properly (build parameters) to the job, and finally works.  this push is just getting rid of some old commented stuff.  we will need to change the url in this file/yml to production from .dev. once migration is complete.